agave-validator: move tests (#7078)
* move verify_args_struct_by_command_run_with_no_genesis_fetch()

* move verify_args_struct_by_command_run_with_no_snapshot_fetch

* move verify_args_struct_by_command_run_with_check_vote_account

* move verify_args_struct_by_command_run_with_only_known_rpc

* move verify_args_struct_by_command_run_with_incremental_snapshot_fetch

* wording
