+++ validator/src/commands/run/args/rpc_bootstrap_config.rs
@@ -51,3 +51,161 @@ impl FromClapArgMatches for RpcBootstrapConfig {
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use {
+        super::*,
+        crate::commands::run::args::{
+            tests::verify_args_struct_by_command_run_with_identity_setup, RunArgs,
+        },
+        solana_pubkey::Pubkey,
+        std::{
+            collections::HashSet,
+            net::{IpAddr, Ipv4Addr, SocketAddr},
+        },
+    };
+
+    #[test]
+    fn verify_args_struct_by_command_run_with_no_genesis_fetch() {
+        // long arg
+        {
+            let default_run_args = RunArgs::default();
+            let expected_args = RunArgs {
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    no_genesis_fetch: true,
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args.clone(),
+                vec!["--no-genesis-fetch"],
+                expected_args,
+            );
+        }
+    }
+
+    #[test]
+    fn verify_args_struct_by_command_run_with_no_snapshot_fetch() {
+        // long arg
+        {
+            let default_run_args = RunArgs::default();
+            let expected_args = RunArgs {
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    no_snapshot_fetch: true,
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args.clone(),
+                vec!["--no-snapshot-fetch"],
+                expected_args,
+            );
+        }
+    }
+
+    #[test]
+    fn verify_args_struct_by_command_run_with_check_vote_account() {
+        // long arg
+        {
+            let default_run_args = RunArgs::default();
+            let expected_args = RunArgs {
+                entrypoints: vec![SocketAddr::new(
+                    IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)),
+                    8000,
+                )],
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    check_vote_account: Some("https://api.mainnet-beta.solana.com".to_string()),
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args,
+                vec![
+                    // required by --check-vote-account
+                    "--entrypoint",
+                    "127.0.0.1:8000",
+                    "--check-vote-account",
+                    "https://api.mainnet-beta.solana.com",
+                ],
+                expected_args,
+            );
+        }
+    }
+
+    #[test]
+    fn verify_args_struct_by_command_run_with_only_known_rpc() {
+        // long arg
+        {
+            let default_run_args = RunArgs::default();
+            let known_validators_pubkey = Pubkey::new_unique();
+            let known_validators = Some(HashSet::from([known_validators_pubkey]));
+            let expected_args = RunArgs {
+                known_validators,
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    only_known_rpc: true,
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args,
+                vec![
+                    // required by --only-known-rpc
+                    "--known-validator",
+                    &known_validators_pubkey.to_string(),
+                    "--only-known-rpc",
+                ],
+                expected_args,
+            );
+        }
+
+        // alias
+        {
+            let default_run_args = RunArgs::default();
+            let known_validators_pubkey = Pubkey::new_unique();
+            let known_validators = Some(HashSet::from([known_validators_pubkey]));
+            let expected_args = RunArgs {
+                known_validators,
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    only_known_rpc: true,
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args,
+                vec![
+                    // required by --no-untrusted-rpc
+                    "--known-validator",
+                    &known_validators_pubkey.to_string(),
+                    "--no-untrusted-rpc",
+                ],
+                expected_args,
+            );
+        }
+    }
+
+    #[test]
+    fn verify_args_struct_by_command_run_with_incremental_snapshot_fetch() {
+        // long arg
+        {
+            let default_run_args = RunArgs::default();
+            let expected_args = RunArgs {
+                rpc_bootstrap_config: RpcBootstrapConfig {
+                    incremental_snapshot_fetch: false,
+                    ..RpcBootstrapConfig::default()
+                },
+                ..default_run_args.clone()
+            };
+            verify_args_struct_by_command_run_with_identity_setup(
+                default_run_args,
+                vec!["--no-incremental-snapshots"],
+                expected_args,
+            );
+        }
+    }
+}
